many practice

+++ Algo/LeadToOffer/48_AddTwoNumber.cpp
@@ -8,7 +8,7 @@ public:
         int carry = 0;
         do {
             sum = num1 ^ num2;
-            carry = (num1 & num2) << 1;
+            carry = (unsigned int) (num1 & num2) << 1;
             num1 = sum;
             num2 = carry;
         } while(num2!=0);
